pyproject.toml
[tool.poetry]
name = "list-of-products"
version = "0.1.0"
description = ""
authors = ["KotOwl <stas.black31@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12"
mysqlclient = ">=2.2.7,<3.0.0"
djangorestframework-simplejwt = ">=5.5.0,<6.0.0"
pillow = ">=11.3.0,<12.0.0"
django-filter = "^25.1"
djangorestframework = "^3.16.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"